api/requirements.txt
tensorflow==2.8.0
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.8.0
matplotlib
numpy